Separate tests for main renderer
